added TaskReader and TaskWriter interface
