--- ToDoly.uml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Diagram>
-  <ID>JAVA</ID>
-  <OriginalElement />
-  <nodes>
-    <node x="412.0" y="494.0">View.View</node>
-    <node x="0.0" y="0.0">Model.Task</node>
-    <node x="0.0" y="494.0">View.Parser</node>
-    <node x="0.0" y="832.0">FIleHandler.TaskReader</node>
-    <node x="432.0" y="0.0">Model.TaskManager</node>
-    <node x="641.0" y="494.0">FIleHandler.TaskWriter</node>
-  </nodes>
-  <notes />
-  <edges />
-  <settings layout="Hierarchic Group" zoom="0.8528536085854203" x="524.0" y="509.9854458606403" />
-  <SelectedNodes />
-  <Categories>
-    <Category>Fields</Category>
-    <Category>Constructors</Category>
-    <Category>Methods</Category>
-    <Category>Inner Classes</Category>
-  </Categories>
-  <SCOPE>All</SCOPE>
-  <VISIBILITY>private</VISIBILITY>
-</Diagram>
-
