+++ src/com/sda/todolist/model/TaskManager.java
@@ -1,4 +1,4 @@
-package com.sda.todolist;
+package com.sda.todolist.model;
 
 import java.time.LocalDate;
 import java.util.*;
@@ -15,12 +15,18 @@ import java.util.Iterator;
 public class TaskManager {
 
     private ArrayList<Task> tasks;
+    private TaskWriter taskWriter;
 
     /**
      * Constructor
      */
-    public TaskManager(ArrayList<Task> tasks) {
-        this.tasks = tasks;
+    public TaskManager(TaskReader taskReader, TaskWriter taskWriter) {
+        this.tasks = taskReader.loadTasks();
+        this.taskWriter = taskWriter;
+    }
+
+    public void saveTasks() {
+        taskWriter.saveTasks(tasks);
     }
 
     /**
