+++ src/com/sda/todolist/model/TaskReader.java
@@ -1,0 +1,8 @@
+package com.sda.todolist.model;
+
+import java.util.ArrayList;
+
+public interface TaskReader {
+
+    ArrayList<Task> loadTasks();
+}
