+++ src/com/sda/todolist/model/TaskWriter.java
@@ -1,0 +1,8 @@
+package com.sda.todolist.model;
+
+import java.util.ArrayList;
+
+public interface TaskWriter {
+
+    void saveTasks(ArrayList<Task> tasks);
+}
